backend/pyproject.toml
[project]
name = "rayolly"
version = "0.1.0"
description = "RayOlly — AI-Native Observability Platform"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.12"
authors = [{ name = "RayOlly Team" }]

dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "uvloop>=0.21.0",
    "httptools>=0.6.0",

    # gRPC for OTLP
    "grpcio>=1.68.0",
    "grpcio-tools>=1.68.0",
    "protobuf>=5.29.0",

    # OpenTelemetry (self-instrumentation + proto definitions)
    "opentelemetry-api>=1.29.0",
    "opentelemetry-sdk>=1.29.0",
    "opentelemetry-proto>=1.29.0",
    "opentelemetry-exporter-otlp>=1.29.0",

    # Storage
    "clickhouse-connect>=0.8.0",
    "duckdb>=1.1.0",
    "pyarrow>=18.0.0",

    # Messaging
    "nats-py>=2.9.0",

    # Cache & state
    "redis[hiredis]>=5.2.0",

    # Database (metadata)
    "asyncpg>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "alembic>=1.14.0",

    # Auth
    "pyjwt[crypto]>=2.10.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.17",

    # AI/ML
    "anthropic>=0.40.0",
    "scikit-learn>=1.6.0",
    "onnxruntime>=1.20.0",
    "prophet>=1.1.6",
    "statsmodels>=0.14.4",

    # Utilities
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "orjson>=3.10.0",
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
    "croniter>=5.0.0",
    "maxminddb>=2.6.0",
    "msgpack>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

gpu = [
    "torch>=2.5.0",
    "vllm>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["rayolly"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH"]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default arg (FastAPI Depends pattern)
    "B904",   # raise from within except (FastAPI HTTPException pattern)
    "B905",   # zip without strict= (Python 3.10+ style preference)
    "B007",   # unused loop control variable
    "SIM101", # merge isinstance calls
    "SIM102", # nested if statements
    "SIM105", # contextlib.suppress vs try/except/pass
    "SIM108", # ternary operator preference
    "E741",   # ambiguous variable name
    "A002",   # argument shadowing builtin
    "N802",   # function name lowercase
    "N803",   # argument name lowercase
    "N815",   # variable in class scope mixedCase
    "TCH002", # move import into TYPE_CHECKING block
    "TC001",  # move import into TYPE_CHECKING block
    "TC002",  # move import into TYPE_CHECKING block
    "TC003",  # move import into TYPE_CHECKING block
    "UP035",  # import from collections.abc
    "UP042",  # str, Enum pattern (StrEnum migration)
    "F841",   # assigned but never used (placeholders)
    "SIM103", # return condition directly
    "SIM117", # combine with statements
]

[tool.ruff.lint.isort]
known-first-party = ["rayolly"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra -q --strict-markers"
markers = [
    "unit: Unit tests",
    "integration: Integration tests requiring external services",
    "e2e: End-to-end tests",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[project.scripts]
rayolly = "rayolly.core.cli:main"
